Rebase the dev_sort_generic branch. Move TreeSortMixin and vars into sort_generic.py
Move TreeSortMixin to sort_generic Part 2.

Fix int problem of datastore filter in sort_tree and add some system tests.

Add another motion.sort system test

Add a circle test to motion sort tests.

Update adapter datastore filter.

The adapter filter function should return Dict[int, PartialModel].
So I moved the code for sort_generic to the adapter and updated the
tests.
