+++ tests/unit/services/test_database_adapter.py
@@ -74,12 +74,12 @@ class DatastoreAdapterTester(TestCase):
         command = commands.Filter(collection=collection, filter=filter)
         self.engine.retrieve.return_value = (
             json.dumps(
-                [
-                    {"f": 1, "meta_deleted": False, "meta_position": 1},
-                    {"f": 1, "meta_deleted": False, "meta_position": 5},
-                    {"f": 1, "meta_deleted": False, "meta_position": 6},
-                    {"f": 1, "meta_deleted": False, "meta_position": 7},
-                ]
+                {
+                    "1": {"f": 1, "meta_deleted": False, "meta_position": 1},
+                    "5": {"f": 1, "meta_deleted": False, "meta_position": 5},
+                    "6": {"f": 1, "meta_deleted": False, "meta_position": 6},
+                    "7": {"f": 1, "meta_deleted": False, "meta_position": 7},
+                }
             ),
             200,
         )
@@ -99,12 +99,12 @@ class DatastoreAdapterTester(TestCase):
         command = commands.Filter(collection=collection, filter=or_filter)
         self.engine.retrieve.return_value = (
             json.dumps(
-                [
-                    {"f": 1, "meta_deleted": False, "meta_position": 1},
-                    {"f": 3, "meta_deleted": False, "meta_position": 4},
-                    {"f": 1, "meta_deleted": False, "meta_position": 6},
-                    {"f": 1, "meta_deleted": False, "meta_position": 7},
-                ]
+                {
+                    "1": {"f": 1, "meta_deleted": False, "meta_position": 1},
+                    "4": {"f": 3, "meta_deleted": False, "meta_position": 4},
+                    "6": {"f": 1, "meta_deleted": False, "meta_position": 6},
+                    "7": {"f": 1, "meta_deleted": False, "meta_position": 7},
+                }
             ),
             200,
         )
